requirements.txt
streamlit
PyMuPDF
langchain
tiktoken
sentence-transformers
google-generativeai
python-dotenv
faiss-cpu
numpy
flask